Use a better merge insertions script
The previous script would not merge insertions that are in the initial
file (the one that fills master_dict) and was not very efficient, since
for each insertion the whole list of insertions is scanned for an
overlap. Instead the input files are concatenated, sorted and clustered
using pybedtools. Clusters are then evaluated to fulfill the same
criteria as before (insertion must overlap and the annotated TE must be
the same to be merged).